LeadConvertFileDiff/requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.4.5.1
chardet==3.0.4
Deprecated==1.2.10
gitdb==4.0.5
GitPython==3.1.11
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
PyGithub==1.53
PyJWT==1.7.1
pylint==2.5.3
pyperclip==1.7.0
requests==2.23.0
six==1.15.0
smmap==3.0.4
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
xmltodict==0.12.0
